requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Flask==2.1.3
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-nmap==0.7.1
requests==2.28.1
urllib3==1.26.10
Werkzeug==2.1.2
